[dev.fuzz] internal/fuzz: use coverage instrumentation while fuzzing
This change updates the go command behavior when
fuzzing to instrument the binary for code coverage,
and uses this coverage in the fuzzing engine to
determine if an input is interesting.

Unfortunately, we can't store and use the coverage
data for a given run of `go test` and re-use it
the next time we fuzz, since the edges could have
changed between builds. Instead, every entry in
the seed corpus and the on-disk corpus is run
by the workers before fuzzing begins, so that the
coordinator can get the baseline coverage for what
the fuzzing engine has already found (or what
the developers have already provided).

Users should run `go clean -fuzzcache` before
using this change, to clear out any existing
"interesting" values that were in the cache.
Previously, every single non-crashing input was
written to the on-disk corpus. Now, only inputs
that actually expand coverage are written.

Bugs that still need fixed:
  - This only works with -parallel=1. There's a
    race when updating the coverage data while
    fuzzing.
  - The testing and internal/fuzz packages are
    instrumented for coverage, which leads to
    more "interesting" values being written to
    the corpus that should be. These packages
    shouldn't be instrumented.

Change-Id: Ie659222d44475c6d68fa4a35d37c37cab3619d71