runtime: fix runtimeQPC ABIInternal call
Currently we call runtimeQPC as ABIInternal because it shaves off 24
bytes by not having an extra wrapper, and at the time we were exceeding
the nosplit stack limit in some cases.

However, this code was written before we had the regabiargs GOEXPERIMENT
flag, and wasn't properly flagged. Naturally, with regabiargs enabled,
it leads to garbage being returned, because it needs to store
runtimeQPC's result to the stack.

Note that now we can no longer tail-call runtimeQPC, because we need to
return back into nanotime1 to set up the return value. This makes us use
8 additional bytes of stack space, which brings us to 8 bytes below the
nosplit limit in the case that was failing that necessitated this
change.

We didn't notice this because today runtimeQPC is only used in Wine, not
on any native Windows platform.

For #40724.

Change-Id: Ia2e5901965ef46c5f299daccef49952026854fe6