cmd/compile: copy definition node for inline vars
clovar creates new ONAME node for local variables or params of closure
inside inlined function, by copying most of the old node fields. So the
new Node.Defn is not modified, its lhs still refer to old node
instead of new one.

That leads to the ICE that ir.StaticValue raises, since when it see an
inlined var but can not find it in the lhs of its definition.

To fix this, we make a separate copy of definition node for clovar, and
modify it to refer to the inlined var instead.

Fixes #45606

Change-Id: Ib517b753a7643756dcd61d36deae60f1a0fc53c5