spec: consolidate the definition of "defined types"
A reader may wrongly think that all defined types are created by type definition.
But this is not true, because types such as bool, int, and string are defined types that are not created by type definition.

And if a reader wants to fully understand what defined types are, the reader has to read several different sections in the spec.

This CL adds a new paragraph to define "defined types" to solve these problems.

Updates #45587

Change-Id: I669d633d6742f8e93bf242fba8d739347e6cceb9