[dev.fuzz] testing,internal/fuzz: prevent unbounded memory growth
Usage of f.testContext.match.fullName to generate the test name causes
unbounded memory growth, eventually causing the fuzzer to slow down
as memory pressure increases.

Each time fuzzFn is invoked it generates a unique string and stores it
in a map. With the fuzzer running at around 100k executions per second
this consumed around ~30GB of memory in a handful of minutes.

Instead calculate and use the SHA256 hash of the input for the test
name, reflecting what the corpus filename would be if a crasher was
found. There is a performance overhead to this, since it requires
calculating the hash after each mutation, but it's cheaper than
running out of memory.

Change-Id: I083f47df7e82f0c6b0bda244f158233784a13029