cmd/go: include packages with InvalidGoFiles when filtering main packages
Empirically, if a package has files that list multiple names go/build
populates the first name encountered and puts the remaining files in
InvalidGoFiles. That foiled our check for packages whose name is
either unpopulated or "main", since the "package main" could be in a
source file after the first.

DO NOT SUBMIT

There is still a likely bug here for literal patterns.

Updates #42088
Fixes #45827

Change-Id: I588314341b17961b38660192c2130678dc03023e