net/url: consider ForceQuery in ResolveReference
Previously, when resolving references of form

  https://golang.org/?hello -> ?

we only used URL.RawQuery to determine whether or not a query part is
defined. Go 1.7 introduced URL.ForceQuery as a flag for the situation
where a query part is provided but empty. But we did not use it in
ResolveReference. This leads to the erroneous output

  https://golang.org/?hello

when the correct output should be

  https://golang.org/?

This commit rectifies that error.

Fixes #46033

Change-Id: I05bc0b48bf2bbf13b4ddc0dd10599ea613dc2188